refactor(php): Run Rector codeQuality

+++ admin/inc/functions_upload.php
@@ -75,7 +75,7 @@ final class functions_upload
         array &$form_errors = []
     ): bool {
         if (! is_array($data) ||
-            empty($data)
+            $data === []
         ) {
             return false;
         }
@@ -89,12 +89,7 @@ final class functions_upload
             }
 
             if (is_bool($upload_form_config[$field]['default'])) {
-                if (isset($value)) {
-                    $value = true;
-                } else {
-                    $value = false;
-                }
-
+                $value = isset($value) ? true : false;
                 $updates[] = [
                     'param' => $field,
                     'value' => functions_mysqli::boolean_to_string($value),
@@ -166,11 +161,7 @@ final class functions_upload
             $original_filename = htmlspecialchars($original_filename);
         }
 
-        if (isset($original_md5sum)) {
-            $md5sum = $original_md5sum;
-        } else {
-            $md5sum = md5_file($source_filepath);
-        }
+        $md5sum = $original_md5sum ?? md5_file($source_filepath);
 
         // we only try to detect duplicate on a new image, not when updating an existing image
         if (! isset($image_id) &&
@@ -291,24 +282,21 @@ final class functions_upload
             $representative_ext = null;
         }
 
-        if (pwg_image::get_library() != 'gd') {
-            if ($conf->original_resize) {
-                $need_resize = self::need_resize($file_path, $conf->original_resize_maxwidth, $conf->original_resize_maxheight);
+        if (pwg_image::get_library() != 'gd' && $conf->original_resize) {
+            $need_resize = self::need_resize($file_path, $conf->original_resize_maxwidth, $conf->original_resize_maxheight);
+            if ($need_resize) {
+                $img = new pwg_image($file_path);
 
-                if ($need_resize) {
-                    $img = new pwg_image($file_path);
+                $img->pwg_resize(
+                    $file_path,
+                    $conf->original_resize_maxwidth,
+                    $conf->original_resize_maxheight,
+                    $conf->original_resize_quality,
+                    $conf->upload_form_automatic_rotation,
+                    false
+                );
 
-                    $img->pwg_resize(
-                        $file_path,
-                        $conf->original_resize_maxwidth,
-                        $conf->original_resize_maxheight,
-                        $conf->original_resize_quality,
-                        $conf->upload_form_automatic_rotation,
-                        false
-                    );
-
-                    $img->destroy();
-                }
+                $img->destroy();
             }
         }
 
@@ -518,7 +506,7 @@ final class functions_upload
             return $representative_ext;
         }
 
-        if (! in_array(strtolower(functions::get_extension($file_path)), ['pdf'])) {
+        if (strtolower(functions::get_extension($file_path)) !== 'pdf') {
             return $representative_ext;
         }
 
@@ -564,7 +552,7 @@ final class functions_upload
             return $representative_ext;
         }
 
-        if (! in_array(strtolower(functions::get_extension($file_path)), ['heic'])) {
+        if (strtolower(functions::get_extension($file_path)) !== 'heic') {
             return $representative_ext;
         }
 
@@ -625,7 +613,7 @@ final class functions_upload
 
         $exec = $conf->ext_imagick_dir . 'convert';
 
-        if ($conf->tiff_representative_ext == 'jpg') {
+        if ($conf->tiff_representative_ext === 'jpg') {
             $exec .= ' -quality 98';
         }
 
@@ -734,7 +722,7 @@ final class functions_upload
             return $representative_ext;
         }
 
-        if (! in_array(strtolower(functions::get_extension($file_path)), ['psd'])) {
+        if (strtolower(functions::get_extension($file_path)) !== 'psd') {
             return $representative_ext;
         }
 
@@ -793,7 +781,7 @@ final class functions_upload
             return $representative_ext;
         }
 
-        if (! in_array(strtolower(functions::get_extension($file_path)), ['eps'])) {
+        if (strtolower(functions::get_extension($file_path)) !== 'eps') {
             return $representative_ext;
         }
 
@@ -856,14 +844,7 @@ final class functions_upload
         // rotation must be applied to the resized photo, then we should test
         // invert width and height.
         [$width, $height] = getimagesize($image_filepath);
-
-        if ($width > $max_width ||
-            $height > $max_height
-        ) {
-            return true;
-        }
-
-        return false;
+        return $width > $max_width || $height > $max_height;
     }
 
     public static function pwg_image_infos(
@@ -884,13 +865,7 @@ final class functions_upload
     ): array {
         global $conf;
 
-        if (isset($conf->upload_form_all_types) &&
-            $conf->upload_form_all_types
-        ) {
-            $extensions = $conf->file_ext;
-        } else {
-            $extensions = $conf->picture_ext;
-        }
+        $extensions = isset($conf->upload_form_all_types) && $conf->upload_form_all_types ? $conf->file_ext : $conf->picture_ext;
 
         return array_unique(array_map(strtolower(...), $extensions));
     }
@@ -932,11 +907,11 @@ final class functions_upload
         $suffix = substr($value, -1);
         $multiply_by = null;
 
-        if ($suffix == 'K') {
+        if ($suffix === 'K') {
             $multiply_by = 1024;
-        } elseif ($suffix == 'M') {
+        } elseif ($suffix === 'M') {
             $multiply_by = 1024 * 1024;
-        } elseif ($suffix == 'G') {
+        } elseif ($suffix === 'G') {
             $multiply_by = 1024 * 1024 * 1024;
         }
 
@@ -968,16 +943,13 @@ final class functions_upload
                     $relative_dir
                 );
             }
-        } else {
+        } elseif (! is_writable($conf->upload_dir)) {
+            chmod($conf->upload_dir, 0777);
             if (! is_writable($conf->upload_dir)) {
-                chmod($conf->upload_dir, 0777);
-
-                if (! is_writable($conf->upload_dir)) {
-                    return sprintf(
-                        functions::l10n('Give write access (chmod 777) to "%s" directory at the root of your Piwigo installation'),
-                        $relative_dir
-                    );
-                }
+                return sprintf(
+                    functions::l10n('Give write access (chmod 777) to "%s" directory at the root of your Piwigo installation'),
+                    $relative_dir
+                );
             }
         }
